requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.3.0
Werkzeug==2.2.2
